plataforma_backup/requirements.txt
django == 4.2.7
crispy-bootstrap5
paramiko
django-environ
psycopg2-binary
python-crontab